use PageVitalsStore in VitalsSummaryChart
